requirements.txt
torch
numpy
Pillow
tqdm
